-#31 - Added item to ui! Item is clickable
- The player can now left-click on an item to add it to the bag
- Item gets removed from UI after adding it to the bag

This will get altered. The left click will only move the player in space,
deselect an item, combine two items if one of the bag is selected or close
the context menu. The functionality itself will persist, it will only get
moved to another place where it gets executed from.