-#52 - Replacing Surfaces with Sprites
-#52 - All UI classes deleted. Started replacing the first UI classes with
sprite-based classes. The context-menu is now completely sprite-based
and working. It prints the button pressed inside the context menu.
Left-click closes the context menu, when clicked outside the context-menu,
while right-click anywhere opens it. The menu always opens inside the
observed display. If the menu is opened in a border area (right border,
bottom border) the menu opens upwards or left to keep it inside the
screen.

TODO: PlayerUI, BagUI